long training

+++ launchers/PBS/TACO/Basketball/MT_pretrained.sh
@@ -16,6 +16,8 @@ WANDB_PROJECT=${WANDB_PROJECT:-"taco_metaworld"}
 FREEZE=${FREEZE:-"false"}
 NO_TACO=${NO_TACO:-"false"}
 AGENT=${AGENT:-"taco"}
+NUM_STEPS=${NUM_STEPS:-220000}
+SAVE_SNAPSHOT=${SAVE_SNAPSHOT:-"false"}
 
 # Set seed argument based on SEED value
 if [ "$SEED" -eq 1 ]; then
@@ -57,7 +59,10 @@ conda activate metataco
 
 
 # Use quotes and escape model path appropriately
-python3 train_metaworld.py agent.pretrained_path=\"${MODEL_PATH}\" exp_name=\"${EXP_NAME}\" seed=$SEED_ARG env_name=$ENV_NAME random_init=$RANDOM_HAND random_goal=$RANDOM_GOAL wandb_tag=$WANDB_TAG wandb_project=$WANDB_PROJECT num_train_frames=220000 agent.freeze_encoder=$FREEZE agent.no_taco=$NO_TACO agent=$AGENT
+python3 train_metaworld.py agent.pretrained_path=\"${MODEL_PATH}\" exp_name=\"${EXP_NAME}\" seed=$SEED_ARG env_name=$ENV_NAME random_init=$RANDOM_HAND random_goal=$RANDOM_GOAL wandb_tag=$WANDB_TAG wandb_project=$WANDB_PROJECT num_train_frames=$NUM_STEPS agent.freeze_encoder=$FREEZE agent.no_taco=$NO_TACO agent=$AGENT save_snapshot=$SAVE_SNAPSHOT
 
-# Cleanup will be triggered automatically by the trap
-rm -rf exp_local/metaworld/$EXP_NAME*
+
+# Cleanup will be triggered automatically by the trap only if save_snapshot is false:
+if [ "$SAVE_SNAPSHOT" = "false" ]; then
+    rm -rf exp_local/metaworld/$EXP_NAME*
+fi
